backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "igorkhaylov"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "psycopg[binary] (>=3.2.9,<4.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "setproctitle (>=1.3.6,<2.0.0)",
    "firebase-admin (>=7.1.0,<8.0.0)",
    "django-rosetta (>=0.10.2,<0.11.0)",
    "django-modeltranslation (>=0.19.17,<0.20.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "django-filter (>=25.1,<26.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "django-storages[s3] (>=1.14.6,<2.0.0)",
    "django-cleanup (>=9.0.0,<10.0.0)",
    "django-stdimage @ git+https://github.com/igorkhaylov/django-stdimage.git@v1.0.0",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
django-debug-toolbar = "^6.0.0"
ruff = "^0.15.4"

